requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
